#4 code reorganization
